FIXED ERROR IN  5-flip_bits.c

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -20,7 +20,7 @@ unsigned int flip_bits(unsigned long int n, unsigned long int m)
  * @n: number with set bits
  * Return: Number of setbits
  */
-unsigned long int count_bits(unsigned long int n)
+unsigned int count_bits(unsigned long int n)
 {
 	int count = 0;
